+++ packages/eslint/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rollup/plugin-eslint ChangeLog
 
+## v8.0.2
+
+_2022-04-11_
+
+### Bugfixes
+
+- fix: filter options should follow createFilter parameter typing (#1140)
+
+### Updates
+
+- chore: upgrade TypeScript (#712)
+
 ## v8.0.1
 
 _2020-11-30_