+++ packages/node-resolve/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rollup/plugin-node-resolve ChangeLog
 
+## v13.2.1
+
+_2022-04-15_
+
+### Bugfixes
+
+- fix: update side effects logic to be deep when glob doesn’t contain `/` (#1148)
+
+## v13.2.0
+
+_2022-04-11_
+
+### Features
+
+- feat: Add the ability to pass a function into resolveOnly (#1152)
+
 ## v13.1.3
 
 _2022-01-05_